Authentication system added
Authentication system added. This covers signing up for an account,
logging in and logging out. base.html and some of its css was changed to
make room for the new login system in the navbar. Finally, accounts was
added as an app to setting.py of the main project, and the urls.py and
views.py were correctly adjusted or created to allow the authentication
system.